+++ docs/devdeploy.jl
@@ -6,8 +6,7 @@ NOTE you have to dev LinDistFlow in the docs environment to get local changes
 function devbuildserve()
     rm("build", force=true, recursive=true)
     include("make.jl")
-    cd("build")
-    serve()
+    serve(dir="build")
 end
 
 devbuildserve()