+++ src/LinDistFlow.jl
@@ -17,8 +17,6 @@ export
     constrain_KVL,
     constrain_loads,
     constrain_line_amps,
-    get_bus_values, 
-    get_edge_values, 
     # recover_voltage_current,  # TODO validate this method
     i_to_j, 
     j_to_k, 
@@ -30,10 +28,12 @@ export
     has_vreg,
     make_graph,
     remove_bus!,
-    combine_parallel_lines!
+    combine_parallel_lines!,
+    Results
 
 include("utils.jl")
 include("model_single_phase.jl")
 include("model_multi_phase.jl")
+include("results.jl")
 
 end # module