+++ src/model_multi_phase.jl
@@ -8,7 +8,7 @@ TODO:
 - currently indexed bus/edge, phase, time
 - but want to align with BranchFlowMmodel MultiPhase: time, bus/edge, phase (broadest to narrowest)
 - and standardize variable containers in CommonOPF
-- but the using nested Dicts makes variable contstruction slow b/c have to create one at time,
+- but using nested Dicts makes variable construction slow b/c have to create one at time,
     (which was necessary in MultiPhase BranchFlowModel) - but maybe not a big deal
 """
 function add_variables(m::JuMP.AbstractModel, p::Inputs{MultiPhase})